feat(api-momo): implemented full CRUD for momo with image upload and tag filtering
- Added Momo entity with Name, Description, Price, ImagePath, and comma-separated Tags
- Implemented endpoints:
  - POST /api/momos to create momo with image upload
  - GET /api/momos to retrieve all momos with optional tag filtering
  - GET /api/momos/{id} to get momo by ID
  - PUT /api/momos/{id} to update momo and handle image replacement
  - DELETE /api/momos/{id} to delete momo and associated image
- Added FileService for saving, deleting, and generating image URLs
- Supported filtering momos by multiple tags in GET endpoint
- Stored tags as comma-separated strings in DB
- Enabled static file serving for uploaded images
- Added validation and graceful error handling

chore(gitignore): excluded image uploads and SQLite database files

+++ Momos.Server/Repositories/Base/Interface/IRepository.cs
@@ -44,10 +44,10 @@ namespace Momos.Server.Repositories.Base.Interface
         Task<List<TEntity>> ListAsync(int page, int pageSize);
 
         // Returns an entity of type TEntity by its primary key from the database.
-        TEntity? Find(long id);
+        TEntity? Find(int id);
 
         // Returns an entity of type TEntity by its primary key from the database asynchronously.
-        Task<TEntity?> FindAsync(long id);
+        Task<TEntity?> FindAsync(int id);
 
         // Returns the DbSet for the specified entity type TEntity from the DbContext.
         DbSet<TEntity> Set();
